rebate

+++ src/main/java/com/project/Enovel/domain/rebate/entity/Rebate.java
@@ -19,7 +19,7 @@ import java.time.LocalDateTime;
 @ToString
 @AllArgsConstructor
 public class Rebate extends BaseEntity {
-    private Long totalPrice;
+    private int totalPrice;
 
     @OneToOne
     private Member member;
